mcp refactor
