+++ mcp_tools/url_recon/waybackurls.py
@@ -1,0 +1,33 @@
+# mcp_tools/url_recon/waybackurls.py
+
+from typing import Dict, Any
+
+def register_waybackurls_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def waybackurls_discovery(domain: str, get_versions: bool = False,
+                             no_subs: bool = False, additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Waybackurls for historical URL discovery with enhanced logging.
+
+        Args:
+            domain: The target domain
+            get_versions: Get all versions of URLs
+            no_subs: Don't include subdomains
+            additional_args: Additional Waybackurls arguments
+
+        Returns:
+            Historical URL discovery results from Wayback Machine
+        """
+        data = {
+            "domain": domain,
+            "get_versions": get_versions,
+            "no_subs": no_subs,
+            "additional_args": additional_args
+        }
+        logger.info(f"🕰️  Starting Waybackurls discovery: {domain}")
+        result = hexstrike_client.safe_post("api/tools/waybackurls", data)
+        if result.get("success"):
+            logger.info(f"✅ Waybackurls discovery completed for {domain}")
+        else:
+            logger.error(f"❌ Waybackurls discovery failed for {domain}")
+        return result
